Changed external return value paradigm
* Tester is compiled separately from those that compile with it, so the
library will not get the symbols that are being appended to
TestReturnValues. Therefore the solution is to dynamically insert new
symbols.
* Added TestExternalValue(const char *, int) that acts as a temporary
external value (which is "ignored" after the next command is parsed)
* .def files removed and testerTools.cmake modified, since they are not
going to be added to projects or to preprocessor definitions
* Added 'ext' optional flag to /catch command for catching external
values (and then, differentiates return values of same VALUE, but one is
NATIVE and the other is EXTERNAL).
* Removed TesterGetExternalValueInfo (or sth like that) external
function
* Fixed several issues in tester.c
* Improved error message (now prints last line read and cursor position)