diff, pragmas and unsigned long -> size_t
* Functions that returned their size by reference (map, set and graph)
as unsigned long now return as size_t, which is semantically more fitting
and may occupy less space (in 32-bit systems, it would occupy 4 bytes, instead
of 8 bytes, which is the size of a long).
* Added token 'z' to testerlib that parses size_t arguments
* Rewritten some tester code to parse size_t, instead of long arguments
* pygraph now returns its size by a separate method, since seq_length returns
Py_ssize_t, which is an opaque type for ssize_t (signed size_t), which can have
overflow issues with size_t. Therefor, another function was created, returning
a PyObject wrapping a size_t value.
* Added pragmas to ignore some MSVC warnings
* Changed diff algorithm and return value